fixed some logic in the timer module which was broken when i switched to seconds instead of ms. removed the EEPROM init code. All it was saving was the do_log? and measurement interval. I will just set these to default values from now on. changed the temperature display code. if there are multiple sensors it now loops between them, displaying them one at a time for 10 seconds each. you can also use the pushbuttons up and down to select which one to display.

+++ MyTimer.h
@@ -30,7 +30,7 @@ public:
 private:
     unsigned long prev_millis; // value returned by the millis() function in the previous run() call
     timer_callback callback;
-    int timer_delay;
+    unsigned long timer_delay;
     boolean enabled; // which timers are enabled
     boolean run_once;
 };
